避免重复下载go-cqhttp, 完善文档

+++ install.sh
@@ -16,16 +16,20 @@ if [ ! -n "$password" ]; then
   exit
 fi
 
-cd go-cqhttp
-wget -O "go-cqhttp.tar.gz" "https://hub.fastgit.org/Mrs4s/go-cqhttp/releases/download/v0.9.22/go-cqhttp-v0.9.22-linux-amd64.tar.gz"
-tar -xzf go-cqhttp.tar.gz
-chmod +x go-cqhttp
-cd ..
+if [ ! -f "go-cqhttp/go-cqhttp" ]; then
+  cd go-cqhttp
+  wget -O "go-cqhttp.tar.gz" "https://hub.fastgit.org/Mrs4s/go-cqhttp/releases/download/v0.9.22/go-cqhttp-v0.9.22-linux-amd64.tar.gz"
+  tar -xzf go-cqhttp.tar.gz
+  chmod +x go-cqhttp
+  cd ..
+fi
 git clone https://github.com/pcrbot/yobot.git
 
 token=`openssl rand -hex 16`
 echo generate token : $token
-cp go-cqhttp/config_template.json go-cqhttp/config.json
+if [ ! -f "go-cqhttp/config.json" ]; then
+  cp go-cqhttp/config_template.json go-cqhttp/config.json
+fi
 sed -i 's/\("uin":\).*/\1\ '"$qq"',/' go-cqhttp/config.json
 sed -i 's/\("password":\).*/\1\ "'"$password"'",/' go-cqhttp/config.json
 sed -i 's/\("access_token":\).*/\1\ "'"$token"'",/' go-cqhttp/config.json
